+++ services/showcase/api/endpoints/routes/line_of_business.py
@@ -23,33 +23,6 @@ router = APIRouter()
 logger = logging.getLogger(__name__)
 
 
-class CredentialsRead(BaseModel):
-    attrs: Dict
-    cred_def_id: str
-    schema_id: str
-
-
-class Credential(BaseModel):
-    cred_type: str
-    cred_protocol: str
-    cred_def_id: str
-    credential: str
-    issue_state: str
-    issue_role: str
-    created_at: str
-    updated_at: str
-    id: str
-
-
-class Workflow(BaseModel):
-    workflow_state: str
-
-
-class CredentialOfferRead(BaseModel):
-    credential: Credential
-    workflow: Workflow | None
-
-
 @router.get(
     "/lobs",
     status_code=status.HTTP_200_OK,
@@ -97,82 +70,6 @@ async def get_out_of_band_messages(
     oob_repo = OutOfBandRepository(db_session=db)
     items = await oob_repo.get_for_lob(lob_id=lob.id)
     return items
-
-
-@router.get(
-    "/lobs/{lob_id}/credential-offer",
-    status_code=status.HTTP_200_OK,
-    response_model=List[CredentialOfferRead],
-)
-async def get_credential_offer(
-    sandbox_id: UUID,
-    lob_id: UUID,
-    db: AsyncSession = Depends(get_db),
-) -> CredentialOfferRead:
-    # make sure lob is in this sandbox...
-    lob_repo = LobRepository(db_session=db)
-    lob = await lob_repo.get_by_id_with_sandbox(sandbox_id, lob_id)
-    # go get all creds the lob holds
-    creds = await traction.tenant_get_credential_offers(lob.wallet_id, lob.wallet_key)
-    return creds
-
-
-@router.post(
-    "/lobs/{lob_id}/credential-offer/{cred_issue_id}/accept",
-    status_code=status.HTTP_200_OK,
-)
-async def accept_credential_offer(
-    sandbox_id: UUID,
-    lob_id: UUID,
-    cred_issue_id: UUID,
-    db: AsyncSession = Depends(get_db),
-):
-    # make sure lob is in this sandbox...
-    lob_repo = LobRepository(db_session=db)
-    lob = await lob_repo.get_by_id_with_sandbox(sandbox_id, lob_id)
-    # go get all creds the lob holds
-    creds = await traction.tenant_accept_credential_offer(
-        lob.wallet_id, lob.wallet_key, cred_issue_id
-    )
-    return creds
-
-
-@router.post(
-    "/lobs/{lob_id}/credential-offer/{cred_issue_id}/reject",
-    status_code=status.HTTP_200_OK,
-)
-async def reject_credential_offer(
-    sandbox_id: UUID,
-    lob_id: UUID,
-    cred_issue_id: UUID,
-    db: AsyncSession = Depends(get_db),
-):
-    # make sure lob is in this sandbox...
-    lob_repo = LobRepository(db_session=db)
-    lob = await lob_repo.get_by_id_with_sandbox(sandbox_id, lob_id)
-    # go get all creds the lob holds
-    creds = await traction.tenant_reject_credential_offer(
-        lob.wallet_id, lob.wallet_key, cred_issue_id
-    )
-    return creds
-
-
-@router.get(
-    "/lobs/{lob_id}/credentials",
-    status_code=status.HTTP_200_OK,
-    response_model=List[CredentialsRead],
-)
-async def get_credentials(
-    sandbox_id: UUID,
-    lob_id: UUID,
-    db: AsyncSession = Depends(get_db),
-) -> CredentialsRead:
-    # make sure lob is in this sandbox...
-    lob_repo = LobRepository(db_session=db)
-    lob = await lob_repo.get_by_id_with_sandbox(sandbox_id, lob_id)
-    # go get all creds the lob holds
-    creds = await traction.tenant_get_credentials(lob.wallet_id, lob.wallet_key)
-    return creds
 
 
 @router.post(
@@ -235,42 +132,6 @@ async def promote_to_issuer(
     return await sandbox.promote_lob_to_issuer(
         sandbox_id=sandbox_id, lob_id=lob_id, db=db
     )
-
-
-@router.post(
-    "/lobs/{lob_id}/students/{student_id}/issue-degree",
-    status_code=status.HTTP_200_OK,
-)
-async def issue_degree(
-    sandbox_id: UUID,
-    lob_id: UUID,
-    student_id: UUID,
-    db: AsyncSession = Depends(get_db),
-) -> None:
-    s_repo = StudentRepository(db_session=db)
-    lob_repo = LobRepository(db_session=db)
-
-    student = await s_repo.get_by_id_in_sandbox(sandbox_id, student_id)
-    faber = await lob_repo.get_by_id_with_sandbox(sandbox_id, lob_id)
-
-    attrs = [
-        {"name": "student_id", "value": student.student_id},
-        {"name": "name", "value": student.name},
-        {"name": "date", "value": student.date.date().strftime("%d-%m-%Y")},
-        {"name": "degree", "value": student.degree},
-        {"name": "age", "value": student.age},
-    ]
-
-    resp = await traction.tenant_issue_credential(
-        faber.wallet_id,
-        faber.wallet_key,
-        str(student.connection_id),
-        alias="degree",
-        cred_def_id=str(faber.cred_def_id),
-        attributes=attrs,
-    )
-
-    return resp
 
 
 @router.post(
